优化告警设置逻辑

+++ database/migrations/2019_09_20_163432_create_module_table.php
@@ -30,7 +30,7 @@ class CreateModuleTable extends Migration
             $table->integer('alarm_per_minute')->default(10)->comment('报警间隔时间(分钟)');
             $table->string('alarm_types', 128)->nullable()->comment('报警类型 1 微信 2 短信 3 邮件');
             $table->tinyInteger('success_rate')->default(0)->comment('成功率阀值(0-100),0表示不开启');
-            $table->tinyInteger('request_total_rate')->default(0)->comment('调用量报警阀值,0表示不开启');
+            $table->integer('request_total_rate')->default(0)->comment('调用量报警阀值,0表示不开启');
             $table->tinyInteger('request_wave_rate')->default(0)->comment('调用量波动阀值(0-100),0表示不开启');
             $table->integer('avg_time_rate')->default(0)->comment('平均耗时报警阀值(ms),0表示不开启');
             $table->tinyInteger('status')->default(1)->comment('状态：1正常 2禁用 -1删除');
